backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
boto3==1.34.130
botocore==1.34.130
psycopg2-binary==2.9.9
python-dotenv==1.0.1
uvicorn==0.30.6
gunicorn==22.0.0
